pipeline/daily_files/requirements.txt
xarray==2023.10.1
numpy==1.26.2
pyyaml
boto3==1.28.84
requests
netcdf4==1.6.5
h5netcdf==1.3.0
geopandas==0.14.4
shapely==2.0.2
python-cmr==0.9.0
s3fs
h5py
scipy